Move Writer<T> into Lithosphere and clean up SyntaxDumper
