+++ Sources/Lithosphere/SyntaxDumper.swift
@@ -8,39 +8,16 @@
 import Foundation
 import Rainbow
 
-public class SyntaxDumper<StreamType: TextOutputStream> {
-  var stream: StreamType
-  var indent = 0
-
-  public init(stream: inout StreamType) {
-    self.stream = stream
-  }
-
-  public func withIndent(_ f: () -> Void) {
-    indent += 2
-    f()
-    indent -= 2
-  }
-
-  public func write(_ string: String) {
-    stream.write(String(repeating: " ", count: indent))
-    stream.write(string)
-  }
-
-  public func line(_ string: String? = nil) {
-    if let string = string { write(string) }
-    stream.write("\n")
-  }
-
+public class SyntaxDumper<StreamType: TextOutputStream>: Writer<StreamType> {
   public func writeLoc(_ loc: SourceLocation?) {
     if let loc = loc {
       let url = URL(fileURLWithPath: loc.file)
-      stream.write(" ")
-      stream.write(url.lastPathComponent.yellow)
-      stream.write(":")
-      stream.write("\(loc.line)".cyan)
-      stream.write(":")
-      stream.write("\(loc.column)".cyan)
+      write(" ")
+      write(url.lastPathComponent.yellow)
+      write(":")
+      write("\(loc.line)".cyan)
+      write(":")
+      write("\(loc.column)".cyan)
     }
   }
 
@@ -50,25 +27,25 @@ public class SyntaxDumper<StreamType: TextOutputStream> {
     case let node as TokenSyntax:
       switch node.tokenKind {
       case .identifier(let name):
-        stream.write("identifier".green.bold)
-        stream.write(" \"\(name)\"".red)
+        write("identifier".green.bold)
+        write(" \"\(name)\"".red)
       default:
-        stream.write("\(node.tokenKind)".green.bold)
+        write("\(node.tokenKind)".green.bold)
       }
       writeLoc(node.startLoc)
     default:
-      stream.write("\(node.raw.kind)".magenta.bold)
+      write("\(node.raw.kind)".magenta.bold)
       writeLoc(node.startLoc)
       withIndent {
         for child in node.children {
-          line()
+          writeLine()
           dump(child, root: false)
         }
       }
     }
-    stream.write(")")
+    write(")")
     if root {
-      line()
+      writeLine()
     }
   }
 }
